+++ apps/realtime/src/app.module.ts
@@ -47,6 +47,7 @@ import { FunctionModule } from './function/function.module';
 import { IntentModule } from './intent/intent.module';
 import { KnowledgeBaseDocumentModule } from './knowledge-base/document/document.module';
 import { IntegrationOauthTokenModule } from './knowledge-base/integration-oauth-token/integration-oauth-token.module';
+import { KnowledgeBaseTagModule } from './knowledge-base/tag/tag.module';
 import { LegacyModule } from './legacy/legacy.module';
 import { MigrationModule } from './migration/migration.module';
 import { createMongoConfig } from './mikro-orm/mongo.config';
@@ -270,6 +271,7 @@ import { KnowledgeBaseSettingsModule } from './knowledge-base/settings/settings.
     IntentModule,
     KnowledgeBaseDocumentModule,
     KnowledgeBaseSettingsModule,
+    KnowledgeBaseTagModule,
     UploadModule,
     ResponseModule,
     AssistantModule,