+++ libs/platform-config/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.60.26](https://github.com/voiceflow/creator-app/compare/@voiceflow/platform-config@1.60.25...@voiceflow/platform-config@1.60.26) (2024-05-06)
+
+**Note:** Version bump only for package @voiceflow/platform-config
+
 ## [1.60.25](https://github.com/voiceflow/creator-app/compare/@voiceflow/platform-config@1.60.24...@voiceflow/platform-config@1.60.25) (2024-05-02)
 
 **Note:** Version bump only for package @voiceflow/platform-config