+++ libs/sdk-logux-designer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.51.16](https://github.com/voiceflow/creator-app/compare/@voiceflow/sdk-logux-designer@2.51.15...@voiceflow/sdk-logux-designer@2.51.16) (2024-05-06)
+
+**Note:** Version bump only for package @voiceflow/sdk-logux-designer
+
+## [2.51.15](https://github.com/voiceflow/creator-app/compare/@voiceflow/sdk-logux-designer@2.51.14...@voiceflow/sdk-logux-designer@2.51.15) (2024-05-06)
+
+**Note:** Version bump only for package @voiceflow/sdk-logux-designer
+
 ## [2.51.14](https://github.com/voiceflow/creator-app/compare/@voiceflow/sdk-logux-designer@2.51.13...@voiceflow/sdk-logux-designer@2.51.14) (2024-05-03)
 
 **Note:** Version bump only for package @voiceflow/sdk-logux-designer